+++ src/common/init.c
@@ -117,6 +117,7 @@ static void defaultConfig(FFinstance* instance)
     initModuleArg(&instance->config.users);
     initModuleArg(&instance->config.bluetooth);
     initModuleArg(&instance->config.sound);
+    initModuleArg(&instance->config.gamepad);
 
     ffStrbufInitA(&instance->config.libPCI, 0);
     ffStrbufInitA(&instance->config.libVulkan, 0);
@@ -385,6 +386,8 @@ static void destroyConfig(FFinstance* instance)
     destroyModuleArg(&instance->config.openCL);
     destroyModuleArg(&instance->config.users);
     destroyModuleArg(&instance->config.bluetooth);
+    destroyModuleArg(&instance->config.sound);
+    destroyModuleArg(&instance->config.gamepad);
 
     ffStrbufDestroy(&instance->config.libPCI);
     ffStrbufDestroy(&instance->config.libVulkan);